requirements.txt
python-telegram-bot==21.6
feedparser==6.0.11
httpx==0.27.2
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.1
